backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
ultralytics==8.1.0
PyYAML==6.0.1
Werkzeug==3.0.1
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.8.1.78
Pillow>=10.0.0
